[Profiler] Memory profiler part 3: Schema parsing and mutable arguments
Pull Request resolved: https://github.com/pytorch/pytorch/pull/86854

The appropriate annotation for a block of memory is a function of time: an input can be mutated in-place to become an activation, a clever kernel might steal the memory of a detached input (such as a mask) to use as output memory, etc.

We could pessimistically assume that all ops mutate all of their inputs, however inspection of schema allows us to significantly narrow that assumption with minimal effort. Checking schemas also allows us to distinguish between dispatcher ops (which have load bearing semantics) and user annotations with reasonably high precision.
ghstack-source-id: 171825582

Differential Revision: [D40220390](https://our.internmc.facebook.com/intern/diff/D40220390/)