Fix comparison in wishart_log_prob
When using SciPy >= 1.7 this test purposely runs into singular samples
which means there are `inf`s in `batched_prop` and `unbatched_prop`.
The difference of 2 `inf`s is `nan` which will fail the `equal(0` check.
However passing the tensors directly to `assertEqual` is not only
supported but the correct way as it will handle `inf` values etc.